Fix up problems
Completed the gs-tar tools.  It might need some better support for the
gscc tool, but is a good start.

Fixed up the gscc tool so it's actually nice to use.

Put in some work into the mc tool.  It's still taking shape.

Unit test framework is working now, and is useful.

Added the initial monitor tool.  It will need to be changed to use the
context library, when it's written.  But it's a good proof of concept.